pyproject.toml
[project]
name = "tsync"
version = "0.1.0"
description = "Sync Tailscale device names and IPs to Cloudflare DNS A records"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["requests>=2.31.0", "python-dotenv>=1.0.0"]
